requirements.txt
flask
flask-cors
gpt4all
elevenlabs
requests
python-dotenv
streamlit
